+++ pcs_test/tier0/lib/commands/resource/test_resource_create.py
@@ -30,7 +30,9 @@ def create(
     allow_invalid_operation=False,
     agent_name="ocf:heartbeat:Dummy",
     allow_invalid_instance_attributes=False,
+    enable_agent_self_validation=False,
 ):
+    # pylint: disable=too-many-arguments
     return resource.create(
         env,
         "A",
@@ -42,6 +44,7 @@ def create(
         ensure_disabled=disabled,
         allow_invalid_operation=allow_invalid_operation,
         allow_invalid_instance_attributes=allow_invalid_instance_attributes,
+        enable_agent_self_validation=enable_agent_self_validation,
     )
 
 
@@ -51,6 +54,7 @@ def create_group(
     disabled=False,
     meta_attributes=None,
     operation_list=None,
+    enable_agent_self_validation=False,
 ):
     return resource.create_in_group(
         env,
@@ -62,6 +66,7 @@ def create_group(
         instance_attributes={},
         wait=wait,
         ensure_disabled=disabled,
+        enable_agent_self_validation=enable_agent_self_validation,
     )
 
 
@@ -75,6 +80,7 @@ def create_clone(
     clone_id=None,
     agent="ocf:heartbeat:Dummy",
     allow_incompatible_clone_meta_attributes=False,
+    enable_agent_self_validation=False,
 ):
     # pylint: disable=too-many-arguments
     return resource.create_as_clone(
@@ -89,6 +95,7 @@ def create_clone(
         wait=wait,
         ensure_disabled=disabled,
         allow_incompatible_clone_meta_attributes=allow_incompatible_clone_meta_attributes,
+        enable_agent_self_validation=enable_agent_self_validation,
     )
 
 
@@ -99,6 +106,7 @@ def create_bundle(
     meta_attributes=None,
     allow_not_accessible_resource=False,
     operation_list=None,
+    enable_agent_self_validation=False,
 ):
     return resource.create_into_bundle(
         env,
@@ -111,6 +119,7 @@ def create_bundle(
         wait=wait,
         ensure_disabled=disabled,
         allow_not_accessible_resource=allow_not_accessible_resource,
+        enable_agent_self_validation=enable_agent_self_validation,
     )
 
 
@@ -395,13 +404,6 @@ class CreateRolesNormalization(TestCase):
             agent_filename=agent_file_name,
         )
         self.config.runner.cib.load(filename=cib_file)
-        self.config.runner.pcmk.resource_agent_self_validation(
-            {},
-            output="",
-            standard="ocf",
-            provider="pacemaker",
-            agent_type="Stateful",
-        )
 
     def create(self, operation_list=None):
         resource.create(
@@ -597,7 +599,7 @@ class Create(TestCase):
     def test_simplest_resource(self):
         self.config.runner.pcmk.load_agent()
         self.config.runner.cib.load()
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
+        # self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(
             resources=fixture_cib_resources_xml_primitive_simplest
         )
@@ -619,7 +621,9 @@ class Create(TestCase):
             returncode=1,
         )
         self.env_assist.assert_raise_library_error(
-            lambda: create(self.env_assist.get_env()),
+            lambda: create(
+                self.env_assist.get_env(), enable_agent_self_validation=True
+            ),
         )
         self.env_assist.assert_reports(
             [
@@ -650,7 +654,9 @@ class Create(TestCase):
             resources=fixture_cib_resources_xml_primitive_simplest
         )
         create(
-            self.env_assist.get_env(), allow_invalid_instance_attributes=True
+            self.env_assist.get_env(),
+            allow_invalid_instance_attributes=True,
+            enable_agent_self_validation=True,
         )
         self.env_assist.assert_reports(
             [
@@ -670,7 +676,10 @@ class Create(TestCase):
             returncode=0,
         )
         self.env_assist.assert_raise_library_error(
-            lambda: create(self.env_assist.get_env()),
+            lambda: create(
+                self.env_assist.get_env(),
+                enable_agent_self_validation=True,
+            ),
         )
         self.env_assist.assert_reports(
             [
@@ -715,7 +724,6 @@ class Create(TestCase):
         )
         self.config.runner.pcmk.load_agent()
         self.config.runner.cib.load()
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(
             resources=fixture_cib_resources_xml_primitive_simplest
         )
@@ -878,7 +886,6 @@ class Create(TestCase):
     def test_resource_with_operation(self):
         self.config.runner.pcmk.load_agent()
         self.config.runner.cib.load()
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(
             resources="""
                 <resources>
@@ -924,14 +931,12 @@ class Create(TestCase):
             ),
         )
         self.config.runner.cib.load()
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(resources=self.fixture_sanitized_operation)
         create(self.env_assist.get_env())
 
     def test_sanitize_operation_id_from_user(self):
         self.config.runner.pcmk.load_agent()
         self.config.runner.cib.load()
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(resources=self.fixture_sanitized_operation)
         create(
             self.env_assist.get_env(),
@@ -1113,9 +1118,6 @@ class Create(TestCase):
                 </resources>
             """,
         )
-        self.config.runner.pcmk.resource_agent_self_validation(
-            dict(state=1), output=""
-        )
         self.config.env.push_cib(
             resources="""
                 <resources>
@@ -1201,7 +1203,6 @@ class Create(TestCase):
         )
         self.config.runner.cib.upgrade()
         self.config.runner.cib.load(filename="cib-empty-3.4.xml")
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(
             resources="""
                 <resources>
@@ -1257,7 +1258,6 @@ class CreateWait(TestCase):
         self.env_assist, self.config = get_env_tools(test_case=self)
         self.config.runner.pcmk.load_agent()
         self.config.runner.cib.load()
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(
             resources=fixture_cib_resources_xml_primitive_simplest,
             wait=TIMEOUT,
@@ -1387,15 +1387,9 @@ class CreateWait(TestCase):
 
 class CreateInGroup(TestCase):
     def setUp(self):
-        self.agent_self_validation_call_name = (
-            "runner.pcmk.resource_agent_self_validation"
-        )
         self.env_assist, self.config = get_env_tools(test_case=self)
         self.config.runner.pcmk.load_agent()
         self.config.runner.cib.load()
-        self.config.runner.pcmk.resource_agent_self_validation(
-            {}, output="", name=self.agent_self_validation_call_name
-        )
 
     def test_simplest_resource(self):
         (
@@ -1438,7 +1432,6 @@ class CreateInGroup(TestCase):
         create_group(self.env_assist.get_env(), wait=False)
 
     def test_cib_upgrade_on_onfail_demote(self):
-        self.config.remove(self.agent_self_validation_call_name)
         self.config.runner.cib.load(
             filename="cib-empty-3.3.xml",
             instead="runner.cib.load",
@@ -1446,7 +1439,6 @@ class CreateInGroup(TestCase):
         )
         self.config.runner.cib.upgrade()
         self.config.runner.cib.load(filename="cib-empty-3.4.xml")
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(
             resources="""
                 <resources>
@@ -1496,6 +1488,34 @@ class CreateInGroup(TestCase):
         )
         self.env_assist.assert_reports(
             [fixture.info(reports.codes.CIB_UPGRADE_SUCCESSFUL)]
+        )
+
+    def test_resource_self_validation_failure(self):
+        self.config.runner.pcmk.resource_agent_self_validation(
+            {},
+            output="""
+            <output source="stderr">not ignored</output>
+            <output source="stdout">this is ignored</output>
+            <output source="stderr">
+            first issue
+            another one
+            </output>
+            """,
+            returncode=1,
+        )
+        self.env_assist.assert_raise_library_error(
+            lambda: create_group(
+                self.env_assist.get_env(), enable_agent_self_validation=True
+            ),
+        )
+        self.env_assist.assert_reports(
+            [
+                fixture.error(
+                    reports.codes.AGENT_SELF_VALIDATION_RESULT,
+                    result="not ignored\nfirst issue\nanother one",
+                    force_code=reports.codes.FORCE,
+                )
+            ]
         )
 
     def test_fail_wait(self):
@@ -1641,15 +1661,9 @@ class CreateInGroup(TestCase):
 
 class CreateAsClone(TestCase):
     def setUp(self):
-        self.agent_self_validation_call_name = (
-            "runner.pcmk.resource_agent_self_validation"
-        )
         self.env_assist, self.config = get_env_tools(test_case=self)
         self.config.runner.pcmk.load_agent()
         self.config.runner.cib.load()
-        self.config.runner.pcmk.resource_agent_self_validation(
-            {}, output="", name=self.agent_self_validation_call_name
-        )
 
     def test_simplest_resource(self):
         (
@@ -1658,6 +1672,34 @@ class CreateAsClone(TestCase):
             )
         )
         create_clone(self.env_assist.get_env(), wait=False)
+
+    def test_resource_self_validation_failure(self):
+        self.config.runner.pcmk.resource_agent_self_validation(
+            {},
+            output="""
+            <output source="stderr">not ignored</output>
+            <output source="stdout">this is ignored</output>
+            <output source="stderr">
+            first issue
+            another one
+            </output>
+            """,
+            returncode=1,
+        )
+        self.env_assist.assert_raise_library_error(
+            lambda: create_clone(
+                self.env_assist.get_env(), enable_agent_self_validation=True
+            ),
+        )
+        self.env_assist.assert_reports(
+            [
+                fixture.error(
+                    reports.codes.AGENT_SELF_VALIDATION_RESULT,
+                    result="not ignored\nfirst issue\nanother one",
+                    force_code=reports.codes.FORCE,
+                )
+            ]
+        )
 
     def test_custom_clone_id(self):
         (
@@ -1670,7 +1712,6 @@ class CreateAsClone(TestCase):
         )
 
     def test_custom_clone_id_error_invalid_id(self):
-        self.config.remove(self.agent_self_validation_call_name)
         self.env_assist.assert_raise_library_error(
             lambda: create_clone(
                 self.env_assist.get_env(), wait=False, clone_id="1invalid"
@@ -1682,7 +1723,6 @@ class CreateAsClone(TestCase):
 
     def test_custom_clone_id_error_id_already_exist(self):
         self.config.remove(name="runner.cib.load")
-        self.config.remove(self.agent_self_validation_call_name)
         self.config.runner.cib.load(
             resources="""
                 <resources>
@@ -1705,7 +1745,6 @@ class CreateAsClone(TestCase):
         self.env_assist.assert_reports([fixture.report_id_already_exist("C")])
 
     def test_cib_upgrade_on_onfail_demote(self):
-        self.config.remove(self.agent_self_validation_call_name)
         self.config.runner.cib.load(
             filename="cib-empty-3.3.xml",
             instead="runner.cib.load",
@@ -1713,7 +1752,6 @@ class CreateAsClone(TestCase):
         )
         self.config.runner.cib.upgrade()
         self.config.runner.cib.load(filename="cib-empty-3.4.xml")
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(
             resources="""<resources>
                 <clone id="A-clone">
@@ -2198,6 +2236,7 @@ class CreateAsCloneFailures(TestCase):
                 agent=agent.full_name,
                 clone_options={"promotable": "1"},
                 allow_incompatible_clone_meta_attributes=True,
+                enable_agent_self_validation=True,
             ),
         )
         self.env_assist.assert_reports(
@@ -2363,7 +2402,6 @@ class CreateInToBundle(TestCase):
         self.config.runner.cib.load(
             filename="cib-empty-3.4.xml", resources=self.fixture_resources_pre
         )
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(
             resources=self.fixture_resource_post_simple_without_network.format(
                 network="""
@@ -2393,13 +2431,11 @@ class CreateInToBundle(TestCase):
 
     def test_simplest_resource(self):
         self.config.runner.cib.load(resources=self.fixture_resources_pre)
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(resources=self.fixture_resources_post_simple)
         create_bundle(self.env_assist.get_env(), wait=False)
 
     def test_bundle_doesnt_exist(self):
         self.config.runner.cib.load(resources=self.fixture_empty_resources)
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.env_assist.assert_raise_library_error(
             lambda: create_bundle(self.env_assist.get_env(), wait=False),
             [
@@ -2422,7 +2458,6 @@ class CreateInToBundle(TestCase):
                     </resources>
                 """
         )
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
 
         self.env_assist.assert_raise_library_error(
             lambda: create_bundle(self.env_assist.get_env(), wait=False),
@@ -2448,7 +2483,6 @@ class CreateInToBundle(TestCase):
                     </resources>
                 """
         )
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.env_assist.assert_raise_library_error(
             lambda: create_bundle(self.env_assist.get_env(), wait=False),
             [
@@ -2463,7 +2497,6 @@ class CreateInToBundle(TestCase):
 
     def test_wait_fail(self):
         self.config.runner.cib.load(resources=self.fixture_resources_pre)
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(
             resources=self.fixture_resources_post_simple,
             wait=TIMEOUT,
@@ -2488,7 +2521,6 @@ class CreateInToBundle(TestCase):
     )
     def test_wait_ok_run_ok(self):
         self.config.runner.cib.load(resources=self.fixture_resources_pre)
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(
             resources=self.fixture_resources_post_simple, wait=TIMEOUT
         )
@@ -2509,7 +2541,6 @@ class CreateInToBundle(TestCase):
     )
     def test_wait_ok_run_fail(self):
         self.config.runner.cib.load(resources=self.fixture_resources_pre)
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(
             resources=self.fixture_resources_post_simple, wait=TIMEOUT
         )
@@ -2534,7 +2565,6 @@ class CreateInToBundle(TestCase):
     )
     def test_disabled_wait_ok_not_running(self):
         self.config.runner.cib.load(resources=self.fixture_resources_pre)
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(
             resources=self.fixture_resources_post_disabled, wait=TIMEOUT
         )
@@ -2553,7 +2583,6 @@ class CreateInToBundle(TestCase):
     )
     def test_disabled_wait_ok_running(self):
         self.config.runner.cib.load(resources=self.fixture_resources_pre)
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(
             resources=self.fixture_resources_post_disabled, wait=TIMEOUT
         )
@@ -2581,7 +2610,6 @@ class CreateInToBundle(TestCase):
                 </resources>
             """
         )
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.env_assist.assert_raise_library_error(
             lambda: create_bundle(self.env_assist.get_env(), wait=False)
         )
@@ -2605,7 +2633,6 @@ class CreateInToBundle(TestCase):
                 </resources>
             """
         )
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(
             resources=(
                 self.fixture_resource_post_simple_without_network.format(
@@ -2638,7 +2665,6 @@ class CreateInToBundle(TestCase):
                 </resources>
             """
         )
-        self.config.runner.pcmk.resource_agent_self_validation({}, output="")
         self.config.env.push_cib(
             resources=(
                 self.fixture_resource_post_simple_without_network.format(
@@ -2654,4 +2680,33 @@ class CreateInToBundle(TestCase):
     def test_ip_range_defined(self):
         self._test_with_network_defined(
             '<network ip-range-start="192.168.100.200"/>'
+        )
+
+    def test_resource_self_validation_failure(self):
+        self.config.runner.cib.load()
+        self.config.runner.pcmk.resource_agent_self_validation(
+            {},
+            output="""
+            <output source="stderr">not ignored</output>
+            <output source="stdout">this is ignored</output>
+            <output source="stderr">
+            first issue
+            another one
+            </output>
+            """,
+            returncode=1,
+        )
+        self.env_assist.assert_raise_library_error(
+            lambda: create_bundle(
+                self.env_assist.get_env(), enable_agent_self_validation=True
+            ),
+        )
+        self.env_assist.assert_reports(
+            [
+                fixture.error(
+                    reports.codes.AGENT_SELF_VALIDATION_RESULT,
+                    result="not ignored\nfirst issue\nanother one",
+                    force_code=reports.codes.FORCE,
+                )
+            ]
         )